+++ apps/backend/package.json
@@ -25,9 +25,14 @@
             "ts-jest": {
                 "tsconfig": "./tsconfig.json"
             },
-            "SOCKET_URL": "http://localhost:8002"
+            "SOCKET_URL": "http://localhost:8002",
+            "CLOSE_TEST_DB": false
         },
-        "globalSetup": "./test/setupGlobal.ts",
+        "globalSetup": "./test/setup.ts",
+        "globalTeardown": "./test/teardown.ts",
+        "setupFilesAfterEnv": [
+            "jest-extended"
+        ],
         "preset": "ts-jest",
         "moduleNameMapper": {
             "~/routes": "<rootDir>/src/routes",
@@ -60,6 +65,7 @@
         "@jest/types": "^29.6.3",
         "@swc/core": "^1.5.5",
         "@swc/jest": "^0.2.36",
+        "@testcontainers/postgresql": "^10.22.0",
         "@tsconfig/node-lts": "^20.1.3",
         "@types/jest": "^29.5.0",
         "@typescript-eslint/eslint-plugin": "^7.7.1",
@@ -75,6 +81,7 @@
         "jest-extended": "^3.2.4",
         "prettier": "^3.2.5",
         "socket.io-client": "^4.7.5",
+        "testcontainers": "^10.22.0",
         "ts-jest": "^29.1.0",
         "ts-node-dev": "^2.0.0",
         "tsconfig-paths": "^4.2.0",
