+++ apps/backend/test/environment.d.ts
@@ -1,0 +1,1 @@
+import 'jest-extended'
