--- apps/backend/test/test.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="jest-extended" />
