+++ apps/backend/tsconfig.json
@@ -1,7 +1,6 @@
 {
     "extends": "@tsconfig/node-lts",
     "include": ["src", "test"],
-    "files": ["test/test.d.ts"],
     "compilerOptions": {
         "declaration": true,
         "removeComments": true,
