+++ package.json
@@ -7,7 +7,8 @@
     "preinstall": "npx only-allow pnpm",
     "dev": "turbo run dev",
     "build": "turbo run build",
-    "start": "turbo run start"
+    "start": "turbo run start",
+    "test:e2e": "turbo run test:e2e"
   },
   "license": "ISC",
   "packageManager": "pnpm@9.0.0",
